+++ packages/components/Anchor/src/index.ts
@@ -1,0 +1,4 @@
+export { default as Anchor } from './Anchor.vue'
+export { default as AnchorLink } from './AnchorLink.vue'
+
+export type * from './types'