+++ packages/lite-ui/src/index.ts
@@ -1,5 +1,6 @@
 export * from '@lite-space/affix'
 export * from '@lite-space/alert'
+export * from '@lite-space/anchor'
 export * from '@lite-space/avatar'
 export * from '@lite-space/badge'
 export * from '@lite-space/breadcrumb'