+++ packages/theme/src/shortcuts/Anchor.ts
@@ -1,0 +1,6 @@
+export const Anchor = {
+  'lt-anchor': '',
+  'lt-anchor-link': 'text-sm transition decoration-none text-gray-400 op-60 hover:op-90',
+  'lt-anchor-link--active': 'text-black/85! op-100!',
+  'lt-anchor-sub-link': 'ml-3'
+}