+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lite-space/theme
 
+## 0.8.0
+
+### Minor Changes
+
+- @lite-space/anchor@0.0.0
+  @lite-space/ui@0.8.0
+  @lite-space/theme@0.8.0
+
 ## 0.7.0
 
 ### Patch Changes